fix(scripts): Optimize switch_to_nftables.sh

+++ center/plugin/nftable/backend/nftable/switch_to_nftables.sh
@@ -21,17 +21,26 @@ enable_nftables() {
 
 # 写入规则文件并加载
 load_idb_nftables_rules() {
-    RULES_FILE="/etc/nftables.idb.nft"
+    RULES_FILE="/etc/nftables.conf"
+
+    if [[ -s "$RULES_FILE" ]]; then
+        echo "Skip: $RULES_FILE already exists and is not empty, not overwriting."
+        return
+    fi
 
     echo "Writing idb-filter rules to $RULES_FILE..."
 
     cat > "$RULES_FILE" <<EOF
 table ip idb-filter {
     chain input {
-        type filter hook input priority 10; policy drop;
+        type filter hook input priority 0; policy drop;
 
         icmp type echo-request drop
         iifname "lo" accept
+        iifname "docker0" accept
+        iifname "br-+" accept
+        iifname "veth+" accept
+        iifname "docker_gwbridge" accept
         ct state established,related accept
         tcp dport 22 accept
         tcp dport 9918 accept
@@ -39,11 +48,11 @@ table ip idb-filter {
     }
 
     chain output {
-        type filter hook output priority 10; policy accept;
+        type filter hook output priority 0; policy accept;
     }
 
     chain forward {
-        type filter hook forward priority 10; policy drop;
+        type filter hook forward priority 0; policy accept;
     }
 }
 EOF
